[CHANGE] Modular providers
