+++ tnnt_discordbot_cogs/cogs/about.py
@@ -22,11 +22,10 @@ from allianceauth.services.hooks import get_extension_logger
 from aadiscordbot.app_settings import get_site_url
 
 # Terra Nanotech Discordbot Cogs
-from tnnt_discordbot_cogs import __title__
 from tnnt_discordbot_cogs.helper import unload_cog
-from tnnt_discordbot_cogs.providers import AppLogger
+from tnnt_discordbot_cogs.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class About(commands.Cog):
