+++ tnnt_discordbot_cogs/cogs/admin.py
@@ -34,11 +34,10 @@ from aadiscordbot.cogs.utils.decorators import sender_is_admin
 from aadiscordbot.utils import auth
 
 # Terra Nanotech Discordbot Cogs
-from tnnt_discordbot_cogs import __title__
 from tnnt_discordbot_cogs.helper import unload_cog
-from tnnt_discordbot_cogs.providers import AppLogger
+from tnnt_discordbot_cogs.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class Admin(commands.Cog):
