+++ tnnt_discordbot_cogs/cogs/auth.py
@@ -23,11 +23,10 @@ from aadiscordbot.app_settings import get_site_url
 from tnnt_templates.app_settings import AppSettings
 
 # Terra Nanotech Discordbot Cogs
-from tnnt_discordbot_cogs import __title__
 from tnnt_discordbot_cogs.helper import unload_cog
-from tnnt_discordbot_cogs.providers import AppLogger
+from tnnt_discordbot_cogs.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class Auth(commands.Cog):
