+++ tnnt_discordbot_cogs/cogs/price_check.py
@@ -20,11 +20,11 @@ from allianceauth.services.hooks import get_extension_logger
 from aadiscordbot import app_settings
 
 # Terra Nanotech Discordbot Cogs
-from tnnt_discordbot_cogs import __title__, __user_agent__
+from tnnt_discordbot_cogs import __user_agent__
 from tnnt_discordbot_cogs.helper import unload_cog
-from tnnt_discordbot_cogs.providers import AppLogger
+from tnnt_discordbot_cogs.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class MarketRegion(Enum):
