+++ tnnt_discordbot_cogs/cogs/welcome.py
@@ -21,12 +21,11 @@ from aadiscordbot.models import WelcomeMessage
 from aadiscordbot.utils.auth import is_user_authenticated
 
 # Terra Nanotech Discordbot Cogs
-from tnnt_discordbot_cogs import __title__
 from tnnt_discordbot_cogs.helper import unload_cog
 from tnnt_discordbot_cogs.models.setting import Setting
-from tnnt_discordbot_cogs.providers import AppLogger
+from tnnt_discordbot_cogs.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 # Discord Bot Settings
 # Channel
